[guppy] move platform-related code into the `platform` module
Also fill out the documentation for the `platform` module.
