+++ guppy/src/graph/summaries.rs
@@ -20,7 +20,6 @@ pub use guppy_summaries::*;
 pub use package_set::*;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeSet;
-pub use target_spec::summaries::PlatformSummary;
 
 /// A type alias for build summaries generated by `guppy`.
 pub type Summary = SummaryWithMetadata<CargoOptionsSummary>;
